executor/cte_test.go: migrate test-infra to testify (#27103)
